Extract file-loading and tui application logic into separate modules
